refactor: open cluster diff in a modal
