--- src/components/organisms/ClusterDiffPane/index.ts
@@ -1,1 +1,0 @@
-export {default} from './ClusterDiffPane';
