+++ src/components/organisms/PaneManager/PaneManager.tsx
@@ -14,7 +14,7 @@ import {
 import Colors, {BackgroundColors} from '@styles/Colors';
 import {AppBorders} from '@styles/Borders';
 import {Row, Col, Content, SplitView} from '@atoms';
-import {ActionsPane, FileTreePane, PluginManagerPane, NavigatorPane, ClustersPane, ClusterDiffPane} from '@organisms';
+import {ActionsPane, FileTreePane, PluginManagerPane, NavigatorPane, ClustersPane} from '@organisms';
 import {LogViewer, GraphView} from '@molecules';
 import featureJson from '@src/feature-flags.json';
 import {
@@ -143,7 +143,6 @@ const PaneManager = () => {
   };
 
   const openClusterDiffDrawer = () => {
-    dispatch(setLeftMenuSelection(''));
     dispatch(openClusterDiff());
   };
 
@@ -219,51 +218,49 @@ const PaneManager = () => {
             )}
           </Space>
         </StyledColumnLeftMenu>
-        {isClusterDiffVisible ? (
-          <ClusterDiffPane />
-        ) : (
-          <StyledColumnPanes style={{width: contentWidth}}>
-            <SplitView
-              contentWidth={contentWidth}
-              left={
-                <>
-                  <div style={{display: leftMenuSelection === 'file-explorer' ? 'inline' : 'none'}}>
-                    <FileTreePane />
-                  </div>
-                  <div
-                    style={{
-                      display:
-                        featureJson.ShowClusterView && leftMenuSelection === 'cluster-explorer' ? 'inline' : 'none',
-                    }}
-                  >
-                    <ClustersPane />
-                  </div>
-                  <div
-                    style={{
-                      display: featureJson.PluginManager && leftMenuSelection === 'plugin-manager' ? 'inline' : 'none',
-                    }}
-                  >
-                    <PluginManagerPane />
-                  </div>
-                </>
-              }
-              hideLeft={!leftActive}
-              nav={<NavigatorPane />}
-              editor={<ActionsPane contentHeight={contentHeight} />}
-              right={
-                <>
-                  {featureJson.ShowGraphView && rightMenuSelection === 'graph' ? (
-                    <GraphView editorHeight={contentHeight} />
-                  ) : undefined}
-                  <div style={{display: rightMenuSelection === 'logs' ? 'inline' : 'none'}}>
-                    <LogViewer editorHeight={contentHeight} />
-                  </div>
-                </>
-              }
-              hideRight={!rightActive}
-            />
-          </StyledColumnPanes>
-        )}
+
+        <StyledColumnPanes style={{width: contentWidth}}>
+          <SplitView
+            contentWidth={contentWidth}
+            left={
+              <>
+                <div style={{display: leftMenuSelection === 'file-explorer' ? 'inline' : 'none'}}>
+                  <FileTreePane />
+                </div>
+                <div
+                  style={{
+                    display:
+                      featureJson.ShowClusterView && leftMenuSelection === 'cluster-explorer' ? 'inline' : 'none',
+                  }}
+                >
+                  <ClustersPane />
+                </div>
+                <div
+                  style={{
+                    display: featureJson.PluginManager && leftMenuSelection === 'plugin-manager' ? 'inline' : 'none',
+                  }}
+                >
+                  <PluginManagerPane />
+                </div>
+              </>
+            }
+            hideLeft={!leftActive}
+            nav={<NavigatorPane />}
+            editor={<ActionsPane contentHeight={contentHeight} />}
+            right={
+              <>
+                {featureJson.ShowGraphView && rightMenuSelection === 'graph' ? (
+                  <GraphView editorHeight={contentHeight} />
+                ) : undefined}
+                <div style={{display: rightMenuSelection === 'logs' ? 'inline' : 'none'}}>
+                  <LogViewer editorHeight={contentHeight} />
+                </div>
+              </>
+            }
+            hideRight={!rightActive}
+          />
+        </StyledColumnPanes>
+
         <StyledColumnRightMenu style={{display: featureJson.ShowRightMenu ? 'inline' : 'none'}}>
           <Space direction="vertical" style={{width: 43}}>
             <Button
